+++ app/lib/common/themes/components/button_theme.dart
@@ -1,7 +1,8 @@
 import 'package:acter/common/themes/colors/color_scheme.dart';
+import 'package:acter/common/themes/components/text_theme.dart';
 import 'package:flutter/material.dart';
 
-ElevatedButtonThemeData elevatedButtonTheme(ColorScheme colors, TextTheme textTheme) =>
+ElevatedButtonThemeData elevatedButtonTheme() =>
     ElevatedButtonThemeData(
       style: ElevatedButton.styleFrom(
         padding: const EdgeInsets.symmetric(
@@ -10,10 +11,10 @@ ElevatedButtonThemeData elevatedButtonTheme(ColorScheme colors, TextTheme textTh
         ),
         elevation: 0,
         textStyle: textTheme.titleMedium?.copyWith(
-          color: colors.primary,
+          color: colorScheme.primary,
           fontSize: 15,
         ),
-        backgroundColor: colors.primary,
+        backgroundColor: colorScheme.primary,
         foregroundColor: Colors.white,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(10),
@@ -21,7 +22,7 @@ ElevatedButtonThemeData elevatedButtonTheme(ColorScheme colors, TextTheme textTh
       ),
     );
 
-ElevatedButtonThemeData dangerButtonThemeMaker(ColorScheme colors, TextTheme textTheme) =>
+ElevatedButtonThemeData dangerButtonThemeMaker() =>
     ElevatedButtonThemeData(
       style: ElevatedButton.styleFrom(
         padding: const EdgeInsets.symmetric(
@@ -30,10 +31,10 @@ ElevatedButtonThemeData dangerButtonThemeMaker(ColorScheme colors, TextTheme tex
         ),
         elevation: 0,
          textStyle: textTheme.titleMedium?.copyWith(
-          color: colors.primary,
+          color: colorScheme.primary,
           fontSize: 15,
         ),
-        backgroundColor: colors.error,
+        backgroundColor: colorScheme.error,
         foregroundColor: Colors.white,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(10),
@@ -41,7 +42,7 @@ ElevatedButtonThemeData dangerButtonThemeMaker(ColorScheme colors, TextTheme tex
       ),
     );
 
-TextButtonThemeData inlineTextButtonThemeMaker(ColorScheme colors) =>
+TextButtonThemeData inlineTextButtonThemeMaker() =>
     TextButtonThemeData(
       style: TextButton.styleFrom(
         foregroundColor: Colors.white,
@@ -52,11 +53,11 @@ TextButtonThemeData inlineTextButtonThemeMaker(ColorScheme colors) =>
       ),
     );
 
-OutlinedButtonThemeData outlinedButtonTheme(ColorScheme colors, TextTheme textTheme) =>
+OutlinedButtonThemeData outlinedButtonTheme() =>
     OutlinedButtonThemeData(
       style: OutlinedButton.styleFrom(
-        foregroundColor: colors.primary,
-        side: BorderSide(color: colors.primary),
+        foregroundColor: colorScheme.primary,
+        side: BorderSide(color: colorScheme.primary),
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(10),
         ),
@@ -65,7 +66,7 @@ OutlinedButtonThemeData outlinedButtonTheme(ColorScheme colors, TextTheme textTh
           vertical: 12,
         ),
         textStyle: textTheme.titleMedium?.copyWith(
-          color: colors.primary,
+          color: colorScheme.primary,
           fontSize: 15,
         ),
         minimumSize: Size.zero,
@@ -73,11 +74,11 @@ OutlinedButtonThemeData outlinedButtonTheme(ColorScheme colors, TextTheme textTh
       ),
     );
 
-TextButtonThemeData textButtonTheme(ColorScheme colors) => TextButtonThemeData(
+TextButtonThemeData textButtonTheme() => TextButtonThemeData(
       style: TextButton.styleFrom(
         elevation: 0,
-        iconColor: colors.onSurface,
-        foregroundColor: colors.onSurface,
+        iconColor: colorScheme.onSurface,
+        foregroundColor: colorScheme.onSurface,
       ),
     );
 final iconButtonTheme = IconButtonThemeData(
